+++ System/Library/Frameworks/ApplicationServices.framework/Versions/A/Frameworks/CoreGraphics.framework/Versions/A/Headers/CGAffineTransform.h
@@ -1,5 +1,5 @@
 /* CoreGraphics - CGAffineTransform.h
- * Copyright (c) 1998-2000 Apple Computer, Inc.
+ * Copyright (c) 1998-2005 Apple Computer, Inc.
  * All rights reserved.
  */
 
@@ -14,8 +14,8 @@ typedef struct CGAffineTransform CGAffineTransform;
 CG_EXTERN_C_BEGIN
 
 struct CGAffineTransform {
-    float a, b, c, d;
-    float tx, ty;
+    CGFloat a, b, c, d;
+    CGFloat tx, ty;
 };
 
 /* The identity transform: [ 1 0 0 1 0 0 ]. */
@@ -24,22 +24,22 @@ CG_EXTERN const CGAffineTransform CGAffineTransformIdentity;
 
 /* Return the transform [ a b c d tx ty ]. */
 
-CG_EXTERN CGAffineTransform CGAffineTransformMake(float a, float b, float c, float d, float tx, float ty);
+CG_EXTERN CGAffineTransform CGAffineTransformMake(CGFloat a, CGFloat b, CGFloat c, CGFloat d, CGFloat tx, CGFloat ty);
 
 /* Return a transform which translates by `(tx, ty)':
  *   t' = [ 1 0 0 1 tx ty ] */
 
-CG_EXTERN CGAffineTransform CGAffineTransformMakeTranslation(float tx, float ty);
+CG_EXTERN CGAffineTransform CGAffineTransformMakeTranslation(CGFloat tx, CGFloat ty);
 
 /* Return a transform which scales by `(sx, sy)':
  *   t' = [ sx 0 0 sy 0 0 ] */
 
-CG_EXTERN CGAffineTransform CGAffineTransformMakeScale(float sx, float sy);
+CG_EXTERN CGAffineTransform CGAffineTransformMakeScale(CGFloat sx, CGFloat sy);
 
 /* Return a transform which rotates by `angle' radians:
  *   t' = [ cos(angle) sin(angle) -sin(angle) cos(angle) 0 0 ] */
 
-CG_EXTERN CGAffineTransform CGAffineTransformMakeRotation(float angle);
+CG_EXTERN CGAffineTransform CGAffineTransformMakeRotation(CGFloat angle);
 
 /* Return true if `t' is the identity transform, false otherwise. */
 
@@ -48,19 +48,19 @@ CG_EXTERN bool CGAffineTransformIsIdentity(CGAffineTransform t) AVAILABLE_MAC_OS
 /* Translate `t' by `(tx, ty)' and return the result:
  *   t' = [ 1 0 0 1 tx ty ] * t */
 
-CG_EXTERN CGAffineTransform CGAffineTransformTranslate(CGAffineTransform t, float tx, float ty);
+CG_EXTERN CGAffineTransform CGAffineTransformTranslate(CGAffineTransform t, CGFloat tx, CGFloat ty);
 
 /* Scale `t' by `(sx, sy)' and return the result:
  *   t' = [ sx 0 0 sy 0 0 ] * t */
 
-CG_EXTERN CGAffineTransform CGAffineTransformScale(CGAffineTransform t, float sx, float sy);
+CG_EXTERN CGAffineTransform CGAffineTransformScale(CGAffineTransform t, CGFloat sx, CGFloat sy);
 
 /* Rotate `t' by `angle' radians and return the result:
  *   t' =  [ cos(angle) sin(angle) -sin(angle) cos(angle) 0 0 ] * t */
 
-CG_EXTERN CGAffineTransform CGAffineTransformRotate(CGAffineTransform t, float angle);
+CG_EXTERN CGAffineTransform CGAffineTransformRotate(CGAffineTransform t, CGFloat angle);
 
-/* Invert `t' and return the result.  If `t' has zero determinant, then `t'
+/* Invert `t' and return the result. If `t' has zero determinant, then `t'
  * is returned unchanged. */
 
 CG_EXTERN CGAffineTransform CGAffineTransformInvert(CGAffineTransform t);
@@ -98,7 +98,8 @@ CG_EXTERN CGRect CGRectApplyAffineTransform(CGRect rect, CGAffineTransform t) AV
 /*** Definitions of inline functions. ***/
 
 CG_INLINE CGAffineTransform
-__CGAffineTransformMake(float a, float b, float c, float d, float tx, float ty)
+__CGAffineTransformMake(CGFloat a, CGFloat b, CGFloat c, CGFloat d,
+    CGFloat tx, CGFloat ty)
 {
     CGAffineTransform t;
 
@@ -113,8 +114,8 @@ __CGPointApplyAffineTransform(CGPoint point, CGAffineTransform t)
 {
     CGPoint p;
 
-    p.x = t.a * point.x + t.c * point.y + t.tx;
-    p.y = t.b * point.x + t.d * point.y + t.ty;
+    p.x = (CGFloat)((double)t.a * point.x + (double)t.c * point.y + t.tx);
+    p.y = (CGFloat)((double)t.b * point.x + (double)t.d * point.y + t.ty);
     return p;
 }
 
@@ -125,8 +126,8 @@ __CGSizeApplyAffineTransform(CGSize size, CGAffineTransform t)
 {
     CGSize s;
 
-    s.width = t.a * size.width + t.c * size.height;
-    s.height = t.b * size.width + t.d * size.height;
+    s.width = (CGFloat)((double)t.a * size.width + (double)t.c * size.height);
+    s.height = (CGFloat)((double)t.b * size.width + (double)t.d * size.height);
     return s;
 }
